Update components to modular approach
