+++ src/app.js
@@ -1,0 +1,5 @@
+import CalculatorContainer from './components/Calculator';
+
+const App = () => (<CalculatorContainer />);
+
+export default App;
