+++ src/components/Display.js
@@ -1,0 +1,7 @@
+const Display = () => (
+  <div className="display">
+    <p className="display">0</p>
+  </div>
+);
+
+export default Display;
